requirements.txt
gunicorn==19.7
python-decouple==3.3
whitenoise==3.3.1
psycopg2==2.8.2
Flask==1.1.2
Flask-RESTful==0.3.8
Flask-Cors==3.0.10
Flask-Migrate==2.7.0
Flask-HTTPAuth==4.4.0
Flask-Mail==0.9.1
passlib==1.7.4
itsdangerous==1.1.0
pytest==6.2.5
Werkzeug~=1.0.1
mysql-connector-python~=8.0.26
requests==2.26.0
numpy~=1.21.4
coverage==6.2
